No longer using HOKU_PROJECT_PATH. Now relying on __FILE__ macro.
In lieu of HOKU_PROJECT_PATH, we are going to allow the use of another environment variable: HOKU_CONFIG_INI. This is **optional**, as opposed to the mandatory project path one before. This also allows one to specify different paths for the nibble and lumberjack databases. Finally, all tests should pass now. No more getting away with *just* three errors.